+++ CHANGELOG.md
@@ -7,7 +7,8 @@
    - `VLCCommandNameStatus`: Retrieve the player status informations
    - `VLCCommandNameStop`: Stops the player to play the current media
    - `VLCCommandNameToggleFullscreen`: Toggles the fullscreen of the player
-   - `VLCCommandNameToggleLoop`: Toggles the loop all medias in the playlist
+   - `VLCCommandNameToggleLoop`: Toggles the loop all items in the playlist
    - `VLCCommandNameTogglePause`: Toggles the pause of the player
    - `VLCCommandNameToggleRandomPlayback`: Toggles random playback
+   - `VLCCommandNameToggleRepeat`: Toggles the repeat of the current item in the plalist
    - `VLCCommandNameVolume`: Changes the volume of the playback